+++ agent-ovs/ovs/test/IntFlowManager_test.cpp
@@ -2534,10 +2534,10 @@ void BaseIntFlowManagerFixture::initExpCon4() {
     PolicyManager::uri_set_t ps, cs;
 
     const opflex::modb::URI& ruleURI_1 = classifier1->getURI();
-    uint32_t clsr1_cookie = intFlowManager.getId(
+    uint64_t clsr1_cookie = intFlowManager.getId(
                          classifier1->getClassId(), ruleURI_1);
     const opflex::modb::URI& ruleURI_2 = classifier2->getURI();
-    uint32_t clsr2_cookie = intFlowManager.getId(
+    uint64_t clsr2_cookie = intFlowManager.getId(
                          classifier2->getClassId(), ruleURI_2);
     ADDF(Bldr(SEND_FLOW_REM).table(POL)
                  .priority(prio)
